+++ application/common/config/main.php
@@ -89,6 +89,8 @@ return [
             'sendMfaOptionRemovedEmails' => Env::get('SEND_MFA_OPTION_REMOVED_EMAILS', true),
             'sendMfaEnabledEmails' => Env::get('SEND_MFA_ENABLED_EMAILS', true),
             'sendMfaDisabledEmails' => Env::get('SEND_MFA_DISABLED_EMAILS', true),
+            'sendMethodReminderEmails' => Env::get('SEND_METHOD_REMINDER_EMAILS', true),
+            'sendMethodPurgedEmails' => Env::get('SEND_METHOD_PURGED_EMAILS', true),
 
             'subjectForInvite' => Env::get('SUBJECT_FOR_INVITE'),
             'subjectForMfaRateLimit' => Env::get('SUBJECT_FOR_MFA_RATE_LIMIT'),
@@ -104,6 +106,8 @@ return [
             'subjectForMfaManager' => Env::get('SUBJECT_FOR_MFA_MANAGER'),
             'subjectForMfaManagerHelp' => Env::get('SUBJECT_FOR_MFA_MANAGER_HELP'),
             'subjectForMethodVerify' => Env::get('SUBJECT_FOR_METHOD_VERIFY'),
+            'subjectForMethodReminder' => Env::get('SUBJECT_FOR_METHOD_REMINDER'),
+            'subjectForMethodPurged' => Env::get('SUBJECT_FOR_METHOD_PURGED'),
 
             'lostSecurityKeyEmailDays' => Env::get('LOST_SECURITY_KEY_EMAIL_DAYS', 62),
             'minimumBackupCodesBeforeNag' => Env::get('MINIMUM_BACKUP_CODES_BEFORE_NAG', 4),
@@ -206,7 +210,7 @@ return [
         ],
         'method' => ArrayHelper::merge(
             [
-                'lifetime' => '+1 day',
+                'lifetime' => '+5 days',
                 'gracePeriod' => '+15 days',
                 'codeLength' => 6,
                 'maxAttempts' => 10,