+++ application/console/controllers/CronController.php
@@ -105,10 +105,20 @@ class CronController extends Controller
         $emailer->sendDelayedMfaRelatedEmails();
     }
 
+    /**
+     * @deprecated
+     */
     public function actionSendMethodVerifyEmails()
     {
         /* @var $emailer Emailer */
         $emailer = \Yii::$app->emailer;
-        $emailer->sendMethodVerifyEmails();
+        $emailer->sendMethodReminderEmails();
+    }
+
+    public function actionSendMethodReminderEmails()
+    {
+        /* @var $emailer Emailer */
+        $emailer = \Yii::$app->emailer;
+        $emailer->sendMethodReminderEmails();
     }
 }